# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Warehouse Item Management API
  description: REST API for managing warehouse items, stock receptions, and location inventory.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
  /api/v1/items:
    post:
      summary: Create a new warehouse item
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseItem'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: SKU already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List warehouse items with filtering and pagination
      operationId: listItems
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseItemList'

  /api/v1/items/categories:
    get:
      summary: Get all distinct categories
      operationId: getCategories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: string

  /api/v1/items/{id}:
    get:
      summary: Get item details by ID
      operationId: getItemById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item details
        '404':
          description: Item not found
    put:
      summary: Update item details
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '200':
          description: Item updated successfully
        '404':
          description: Item not found
        '409':
          description: SKU already exists
    delete:
      summary: Soft delete item
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item soft deleted successfully
        '404':
          description: Item not found

  /api/v1/stock/receive:
    post:
      summary: Receive stock into location
      operationId: receiveStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiveStockRequest'
      responses:
        '200':
          description: Stock received successfully
        '400':
          description: Invalid quantity or validation error
        '404':
          description: Item or Location not found

  /api/v1/stock/transfer:
    post:
      summary: Transfer stock between warehouse locations
      operationId: transferStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferStockRequest'
      responses:
        '200':
          description: Stock transferred successfully
        '400':
          description: Invalid quantity, same source/destination, or insufficient stock
        '404':
          description: Item or Location not found

  /api/v1/stock/{item_id}:
    get:
      summary: Get stock details per location for an item
      operationId: getStockByItemId
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock list per location

  /api/v1/stock/{item_id}/logs:
    get:
      summary: Get stock mutation logs history for an item
      operationId: getStockLogsByItemId
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock mutation logs history

  /api/v1/locations:
    get:
      summary: List all warehouse locations
      operationId: listLocations
      responses:
        '200':
          description: List of warehouse locations

components:
  schemas:
    CreateItemRequest:
      type: object
      required:
        - sku
        - name
        - category
        - unit
      properties:
        sku:
          type: string
          example: "ELEC-001"
        name:
          type: string
          example: "Wireless Mouse"
        category:
          type: string
          example: "Electronics"
        unit:
          type: string
          example: "pcs"
    ReceiveStockRequest:
      type: object
      required:
        - item_id
        - location_id
        - qty
      properties:
        item_id:
          type: string
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        location_id:
          type: string
          example: "a11ac10b-58cc-4372-a567-0e02b2c3d999"
        qty:
          type: integer
          example: 50
    TransferStockRequest:
      type: object
      required:
        - item_id
        - from_location_id
        - to_location_id
        - qty
      properties:
        item_id:
          type: string
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        from_location_id:
          type: string
          example: "a11ac10b-58cc-4372-a567-0e02b2c3d999"
        to_location_id:
          type: string
          example: "b22bc20c-58cc-4372-a567-0e02b2c3d888"
        qty:
          type: integer
          example: 10
    SuccessResponseItem:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Item created successfully"
        data:
          type: object
    SuccessResponseItemList:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Items retrieved successfully"
        data:
          type: array
          items:
            type: object
        meta:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 42
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "SKU already exists"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "sku"
              reason:
                type: string
                example: "Duplicate entry"